Move video plugin data types in shared lib folder
