+++ shared/lib/video.ts
@@ -1,26 +1,44 @@
-import { parseConfigUUIDs } from './config'
+import {parseConfigUUIDs} from './config'
+import {VideoObject} from '@peertube/peertube-types';
+import {ActivityPubAttributedTo} from '@peertube/peertube-types/shared/models/activitypub/objects/common-objects';
+
+interface ShigVideoObject extends VideoObject {
+    guests: ActivityPubAttributedTo[]
+}
 
 interface VideoHasShigSettings {
-  'shig-per-live-video': boolean
-  'shig-all-lives': boolean
-  'shig-videos-list': string
+    'shig-per-live-video': boolean
+    'shig-all-lives': boolean
+    'shig-videos-list': string
 }
 
-interface SharedVideoBase {
-  uuid: string
-  isLive: boolean
-  pluginData?: {
-    'shigActive'?: boolean
-    'shig-remote'?: boolean
-  }
+interface ShigPluginData {
+    url: string
+
+    shigActive?: boolean
+    shigRemote?: boolean
+
+    firstGuest?: string
+    secondGuest?: string
+    thirdGuest?: string
+
+    customMessage?: string
 }
 
-interface SharedVideoFrontend extends SharedVideoBase {
-  isLocal: boolean
+interface ShigVideo {
+    id: number | string
+    uuid: string
+    url: string
+    isLive: boolean
+    pluginData?: ShigPluginData
 }
 
-interface SharedVideoBackend extends SharedVideoBase {
-  remote: boolean
+interface SharedVideoFrontend extends ShigVideo {
+    isLocal: boolean
+}
+
+interface SharedVideoBackend extends ShigVideo {
+    remote: boolean
 }
 
 type SharedVideo = SharedVideoBackend | SharedVideoFrontend
@@ -31,32 +49,33 @@ type SharedVideo = SharedVideoBackend | SharedVideoFrontend
  * @param video the video
  * @returns true if the video has a local chat
  */
-function videoHasShig (settings: VideoHasShigSettings, video: SharedVideo): boolean {
-  // Never use webchat on remote videos.
-  if ('isLocal' in video) {
-    if (!video.isLocal) return false
-  } else {
-    if (video.remote) return false
-  }
+function videoHasShig(settings: VideoHasShigSettings, video: SharedVideo): boolean {
+    // Never use webchat on remote videos.
+    if ('isLocal' in video) {
+        if (!video.isLocal) return false
+    } else {
+        if (video.remote) return false
+    }
 
-  if (settings['shig-per-live-video'] && video.isLive && video.pluginData && video.pluginData['shigActive']) {
-    return true
-  }
+    if (settings['shig-per-live-video'] && video.isLive && video.pluginData && video.pluginData['shigActive']) {
+        return true
+    }
 
-  if (settings['shig-all-lives']) {
-    if (video.isLive) return true
-  }
+    if (settings['shig-all-lives']) {
+        if (video.isLive) return true
+    }
 
-  const uuids = parseConfigUUIDs(settings['shig-videos-list'])
-  if (uuids.includes(video.uuid)) {
-    return true
-  }
+    const uuids = parseConfigUUIDs(settings['shig-videos-list'])
 
-  return false
+    if (uuids.includes(video.uuid)) {
+        return true
+    }
+
+    return false
 }
 
 interface VideoHasRemoteShigSettings {
-  'federation-no-remote-chat': boolean
+    'federation-no-remote-chat': boolean
 }
 
 /**
@@ -65,19 +84,31 @@ interface VideoHasRemoteShigSettings {
  * @param video the video
  * @returns true if the video has a remote chat
  */
-function videoHasRemoteShig (settings: VideoHasRemoteShigSettings, video: SharedVideo): boolean {
-  if (settings['federation-no-remote-chat']) { return false }
-  if ('isLocal' in video) {
-    if (video.isLocal) return false
-  } else {
-    if (!video.remote) return false
-  }
-  if (!video.pluginData) { return false }
-  if (!video.pluginData['shig-remote']) { return false }
-  return true
+function videoHasRemoteShig(settings: VideoHasRemoteShigSettings, video: SharedVideo): boolean {
+    if (settings['federation-no-remote-chat']) {
+        return false
+    }
+    if ('isLocal' in video) {
+        if (video.isLocal) return false
+    } else {
+        if (!video.remote) return false
+    }
+    if (!video.pluginData) {
+        return false
+    }
+    if (!video.pluginData['shigRemote']) {
+        return false
+    }
+    return true
+}
+
+export type {
+    ShigVideo,
+    ShigPluginData,
+    ShigVideoObject,
 }
 
 export {
-  videoHasShig,
-  videoHasRemoteShig
+    videoHasShig,
+    videoHasRemoteShig
 }
